third_party/WebKit/LayoutTests/media/track/regions-webvtt/vtt-region-parser.html

<!DOCTYPE html>
<title>Tests proper parsing of various regions present in WebVTT header area.</title>
<script src="../../media-file.js"></script>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var video = document.createElement('video');
    video.src = findMediaFile('video', '../../content/test');
    var testTrack = document.createElement('track');
    testTrack.onload = t.step_func_done(function() {
        var track = testTrack.track;
        assert_equals(track.regions.length, 5);

        var region = track.regions[0];
        assert_equals(region.id, 'region_without_settings');

        region = track.regions[1];
        assert_equals(region.id, 'region_with_all_settings');
        assert_equals(region.width, 32);
        assert_equals(region.height, 5);
        assert_equals(region.regionAnchorX, 41);
        assert_equals(region.regionAnchorY, 20);
        assert_equals(region.viewportAnchorX, 31);
        assert_equals(region.viewportAnchorY, 84);
        assert_equals(region.scroll, 'up');

        region = track.regions[2];
        assert_equals(region.id, 'region_floating_point_anchor');
        assert_equals(Math.round(region.regionAnchorX * 1000), 41133);
        assert_equals(Math.round(region.regionAnchorY * 1000), 20420);
        assert_equals(Math.round(region.viewportAnchorX * 1000), 32330);
        assert_equals(Math.round(region.viewportAnchorY * 1000), 32440);

        region = track.regions[3];
        assert_equals(region.id, 'not_unique_id');
        assert_equals(region.width, 67);

        region = track.regions[4];
        assert_equals(region.id, '');
    });
    testTrack.src = '../captions-webvtt/header-regions.vtt';
    testTrack.kind = 'captions';
    testTrack.default = true;
    video.appendChild(testTrack);
});
</script>
